move chan_open_confirm
