+++ crates/ibc/src/core/context.rs
@@ -1,5 +1,6 @@
 mod chan_close_confirm;
 mod chan_close_init;
+mod chan_open_confirm;
 
 mod acknowledgement;
 mod recv_packet;
@@ -9,6 +10,7 @@ use crate::prelude::*;
 
 use self::chan_close_confirm::{chan_close_confirm_execute, chan_close_confirm_validate};
 use self::chan_close_init::{chan_close_init_execute, chan_close_init_validate};
+use self::chan_open_confirm::{chan_open_confirm_execute, chan_open_confirm_validate};
 
 use self::acknowledgement::{acknowledgement_packet_execute, acknowledgement_packet_validate};
 use self::recv_packet::{recv_packet_execute, recv_packet_validate};
@@ -33,13 +35,10 @@ use crate::core::ics03_connection::version::{
 use crate::core::ics04_channel::channel::{ChannelEnd, Counterparty, State};
 use crate::core::ics04_channel::commitment::{AcknowledgementCommitment, PacketCommitment};
 use crate::core::ics04_channel::context::calculate_block_delay;
-use crate::core::ics04_channel::events::{OpenAck, OpenConfirm, OpenInit, OpenTry};
-use crate::core::ics04_channel::handler::{
-    chan_open_ack, chan_open_confirm, chan_open_init, chan_open_try,
-};
+use crate::core::ics04_channel::events::{OpenAck, OpenInit, OpenTry};
+use crate::core::ics04_channel::handler::{chan_open_ack, chan_open_init, chan_open_try};
 use crate::core::ics04_channel::msgs::acknowledgement::Acknowledgement;
 use crate::core::ics04_channel::msgs::chan_open_ack::MsgChannelOpenAck;
-use crate::core::ics04_channel::msgs::chan_open_confirm::MsgChannelOpenConfirm;
 use crate::core::ics04_channel::msgs::chan_open_init::MsgChannelOpenInit;
 use crate::core::ics04_channel::msgs::chan_open_try::MsgChannelOpenTry;
 use crate::core::ics04_channel::msgs::{ChannelMsg, PacketMsg};
@@ -898,90 +897,6 @@ where
 
         for log_message in extras.log {
             ctx_a.log_message(log_message);
-        }
-    }
-
-    Ok(())
-}
-
-fn chan_open_confirm_validate<ValCtx>(
-    ctx_b: &ValCtx,
-    module_id: ModuleId,
-    msg: MsgChannelOpenConfirm,
-) -> Result<(), ContextError>
-where
-    ValCtx: ValidationContext,
-{
-    chan_open_confirm::validate(ctx_b, &msg)?;
-
-    let module = ctx_b
-        .get_route(&module_id)
-        .ok_or(ChannelError::RouteNotFound)?;
-    module.on_chan_open_confirm_validate(&msg.port_id_on_b, &msg.chan_id_on_b)?;
-
-    Ok(())
-}
-
-fn chan_open_confirm_execute<ExecCtx>(
-    ctx_b: &mut ExecCtx,
-    module_id: ModuleId,
-    msg: MsgChannelOpenConfirm,
-) -> Result<(), ContextError>
-where
-    ExecCtx: ExecutionContext,
-{
-    let module = ctx_b
-        .get_route_mut(&module_id)
-        .ok_or(ChannelError::RouteNotFound)?;
-
-    let extras = module.on_chan_open_confirm_execute(&msg.port_id_on_b, &msg.chan_id_on_b)?;
-
-    let chan_end_on_b = ctx_b.channel_end(&(msg.port_id_on_b.clone(), msg.chan_id_on_b.clone()))?;
-
-    // state changes
-    {
-        let port_channel_id_on_b = (msg.port_id_on_b.clone(), msg.chan_id_on_b.clone());
-        let chan_end_on_b = {
-            let mut chan_end_on_b = chan_end_on_b.clone();
-            chan_end_on_b.set_state(State::Open);
-
-            chan_end_on_b
-        };
-
-        ctx_b.store_channel(port_channel_id_on_b, chan_end_on_b)?;
-    }
-
-    // emit events and logs
-    {
-        ctx_b.log_message("success: channel open confirm".to_string());
-
-        let conn_id_on_b = chan_end_on_b.connection_hops[0].clone();
-        let port_id_on_a = chan_end_on_b.counterparty().port_id.clone();
-        let chan_id_on_a = chan_end_on_b
-            .counterparty()
-            .channel_id
-            .clone()
-            .ok_or(ContextError::ChannelError(ChannelError::Other {
-            description:
-                "internal error: ChannelEnd doesn't have a counterparty channel id in OpenConfirm"
-                    .to_string(),
-        }))?;
-
-        let core_event = IbcEvent::OpenConfirmChannel(OpenConfirm::new(
-            msg.port_id_on_b.clone(),
-            msg.chan_id_on_b.clone(),
-            port_id_on_a,
-            chan_id_on_a,
-            conn_id_on_b,
-        ));
-        ctx_b.emit_ibc_event(core_event);
-
-        for module_event in extras.events {
-            ctx_b.emit_ibc_event(IbcEvent::AppModule(module_event));
-        }
-
-        for log_message in extras.log {
-            ctx_b.log_message(log_message);
         }
     }
